Updated to 64bit version

+++ module/pna_main.c
@@ -61,8 +61,8 @@ struct pna_perf {
     __u64 t_jiffies; /* 8 */
     struct timeval currtime; /* 8 */
     struct timeval prevtime; /* 8 */
-    __u32 p_interval[PNA_DIRECTIONS]; /* 8 */
-    __u32 B_interval[PNA_DIRECTIONS]; /* 8 */
+    __u64 p_interval[PNA_DIRECTIONS]; /* 16 */
+    __u64 B_interval[PNA_DIRECTIONS]; /* 16 */
     pna_stat_uword dev_last_rx;
     pna_stat_uword dev_last_fifo;
     unsigned int contention_miss;
@@ -294,9 +294,9 @@ static void pna_nslog(struct timespec *start, struct timespec *stop, struct pna_
  */
 static void pna_perflog(struct sk_buff *skb, int dir, struct net_device *dev)
 {
-    __u32 t_interval;
-    __u32 fps_in, Mbps_in, avg_in;
-    __u32 fps_out, Mbps_out, avg_out;
+    __u64 t_interval;
+    __u64 fps_in, Mbps_in, avg_in;
+    __u64 fps_out, Mbps_out, avg_out;
     pna_link_stats stats;
     __u32 frame_count;
     struct pna_perf *perf = &get_cpu_var(perf_data);
@@ -337,8 +337,8 @@ static void pna_perflog(struct sk_buff *skb, int dir, struct net_device *dev)
         /* report the numbers */
         if (fps_in + fps_out > 1000) {
             pr_info("pna flowmon_smpid:%d,contention:%u,"
-                    "in_fps:%u,in_Mbps:%u,in_avg:%u,"
-                    "out_fps:%u,out_Mbps:%u,out_avg:%u\n", smp_processor_id(),
+                    "in_fps:%llu,in_Mbps:%llu,in_avg:%llu,"
+                    "out_fps:%llu,out_Mbps:%llu,out_avg:%llu\n", smp_processor_id(),
                     perf->contention_miss,
                     fps_in, Mbps_in, avg_in, fps_out, Mbps_out, avg_out);
